Move NodeFile into own module and start to add functions similar to NodeBlock to NodeFile
